requirements.txt
asgiref==3.8.1
Django==5.0.7
django-allauth==0.63.6
django-cors-headers==4.4.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-decouple==3.8
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
